CLI and words filter added
