+++ TagCloud.Common/Drawing/ICloudDrawer.cs
@@ -1,0 +1,8 @@
+﻿using TagCloud.Common.Options;
+
+namespace TagCloud.Common.Drawing;
+
+public interface ICloudDrawer
+{
+    void DrawCloud(IEnumerable<Tag> tags, VisualizationOptions visualizationOptions);
+}
