+++ TagCloud.Common/Layouter/ICloudLayouter.cs
@@ -1,0 +1,10 @@
+﻿using System.Drawing;
+
+namespace TagCloud.Common.Layouter;
+
+public interface ICloudLayouter
+{
+    Rectangle PutNextRectangle(Size rectangleSize);
+    IEnumerable<Rectangle> GetRectanglesLayout();
+    void ClearRectanglesLayout();
+}
