+++ TagCloud.Common/Layouter/IncorrectSizeException.cs
@@ -1,0 +1,8 @@
+﻿namespace TagCloud.Common.Layouter;
+
+public class IncorrectSizeException : Exception
+{
+    public IncorrectSizeException() : base("Size of rectangle must be positive")
+    {
+    }
+}
