+++ TagCloud.Common/TextFilter/ITextFilter.cs
@@ -1,0 +1,6 @@
+﻿namespace TagCloud.Common.TextFilter;
+
+public interface ITextFilter
+{
+    public IEnumerable<string> FilterAllWords(string pathToFile, int boringWordsLength);
+}
