+++ TagCloud.Common/WeightCounter/IWeightCounter.cs
@@ -1,0 +1,6 @@
+﻿namespace TagCloud.Common.WeightCounter;
+
+public interface IWeightCounter
+{
+    Dictionary<string, int> CountWeights(IEnumerable<string> lines);
+}
