+++ TagCloud.Common/WeightCounter/SimpleWeightCounter.cs
@@ -1,0 +1,13 @@
+﻿namespace TagCloud.Common.WeightCounter;
+
+public class SimpleWeightCounter : IWeightCounter
+{
+    public Dictionary<string, int> CountWeights(IEnumerable<string> lines)
+    {
+        return lines
+            .GroupBy(x => x)
+            .ToDictionary(x => x.Key, x => x.Count())
+            .OrderByDescending(x => x.Value)
+            .ToDictionary(x => x.Key, x => x.Value);
+    }
+}
