+++ TagCloud.GUI/MainForm.cs
@@ -1,0 +1,9 @@
+namespace TagCloud.GUI;
+
+public sealed class MainForm : Form
+{
+    public MainForm()
+    {
+      
+    }
+}
