+++ TagCloud.GUI/Program.cs
@@ -1,0 +1,11 @@
+namespace TagCloud.GUI;
+
+internal static class Program
+{
+    [STAThread]
+    private static void Main()
+    {
+        ApplicationConfiguration.Initialize();
+        Application.Run(new MainForm());
+    }
+}
