--- TagsCloudContainer/Drawing/ICloudDrawer.cs
@@ -1,8 +1,0 @@
-﻿using TagsCloudContainer.Options;
-
-namespace TagsCloudContainer.Drawing;
-
-public interface ICloudDrawer
-{
-    void DrawCloud(IEnumerable<Tag> tags, VisualizationOptions visualizationOptions);
-}
