--- TagsCloudContainer/Layouter/IncorrectSizeException.cs
@@ -1,8 +1,0 @@
-﻿namespace TagsCloudContainer.Layouter;
-
-public class IncorrectSizeException : Exception
-{
-    public IncorrectSizeException() : base("Size of rectangle must be positive")
-    {
-    }
-}
