--- TagsCloudContainer/TagsConverter/ITagsConverter.cs
@@ -1,8 +1,0 @@
-﻿using TagsCloudContainer.Options;
-
-namespace TagsCloudContainer.TagsConverter;
-
-public interface ITagsConverter
-{
-    IEnumerable<Tag> ConvertToTags(IEnumerable<string> words, VisualizationOptions visualizationOptions);
-}
