--- TagsCloudContainer/WeightCounter/IWeightCounter.cs
@@ -1,6 +1,0 @@
-﻿namespace TagsCloudContainer.WeightCounter;
-
-public interface IWeightCounter
-{
-    Dictionary<string, int> CountWeights(IEnumerable<string> lines);
-}
