+++ di.sln
@@ -2,9 +2,11 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FractalPainter", "FractalPainter\FractalPainter.csproj", "{4D70883B-6F8B-4166-802F-8EDC9BE93199}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagsCloudContainer", "TagsCloudContainer\TagsCloudContainer.csproj", "{086DF112-6CA4-4882-9C66-8673276B2CBE}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagCloud.Tests", "TagCloud.Tests\TagCloud.Tests.csproj", "{B06D2F37-D75C-448D-900C-CD4CA7D25952}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagsCloudContainer.Tests", "TagsCloudContainer.Tests\TagsCloudContainer.Tests.csproj", "{B06D2F37-D75C-448D-900C-CD4CA7D25952}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagCloud.Common", "TagCloud.Common\TagCloud.Common.csproj", "{B417C383-0A23-4043-B4DC-8F72D32167FF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagCloud.CLI", "TagCloud.CLI\TagCloud.CLI.csproj", "{8DDC37BF-3171-4F99-8B62-3E50B82E7272}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -16,13 +18,17 @@ Global
 		{4D70883B-6F8B-4166-802F-8EDC9BE93199}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4D70883B-6F8B-4166-802F-8EDC9BE93199}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4D70883B-6F8B-4166-802F-8EDC9BE93199}.Release|Any CPU.Build.0 = Release|Any CPU
-		{086DF112-6CA4-4882-9C66-8673276B2CBE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{086DF112-6CA4-4882-9C66-8673276B2CBE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{086DF112-6CA4-4882-9C66-8673276B2CBE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{086DF112-6CA4-4882-9C66-8673276B2CBE}.Release|Any CPU.Build.0 = Release|Any CPU
 		{B06D2F37-D75C-448D-900C-CD4CA7D25952}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{B06D2F37-D75C-448D-900C-CD4CA7D25952}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B06D2F37-D75C-448D-900C-CD4CA7D25952}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B06D2F37-D75C-448D-900C-CD4CA7D25952}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B417C383-0A23-4043-B4DC-8F72D32167FF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B417C383-0A23-4043-B4DC-8F72D32167FF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B417C383-0A23-4043-B4DC-8F72D32167FF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B417C383-0A23-4043-B4DC-8F72D32167FF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8DDC37BF-3171-4F99-8B62-3E50B82E7272}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8DDC37BF-3171-4F99-8B62-3E50B82E7272}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8DDC37BF-3171-4F99-8B62-3E50B82E7272}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8DDC37BF-3171-4F99-8B62-3E50B82E7272}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
